Move test_views to example apps
